requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
psycopg2-binary==2.9.9
pandas>=2.0.0
xlsxwriter>=3.1.0
openpyxl>=3.1.0
pytest>=8.0.0
pytest-cov>=4.1.0
selenium>=4.10.0
Werkzeug==2.3.7
email-validator==2.0.0
python-dotenv==1.0.0
boto3>=1.34.0
supabase==2.0.0
requests>=2.31.0